feat: Portal components with isolated sdk styles
